Simplify paths
Reduce indirection in file resolution. This allows paths expressed in
the source code to more directly relate to the true location of files on
disk. It also increases parity in how paths are interpreted during
development and when building the application for production.

The changes to "assets" paths were generated with the following command:

    $ git ls-files src/assets src/modules src/pages | xargs sed -i 's/\(\.\.\/\)\?assets\//src\/assets\//g'